+++ reana_commons/config.py
@@ -353,8 +353,13 @@ HTCONDOR_JOB_FLAVOURS = {
     "testmatch": 259200,
     "nextweek": 604800,
 }
-
 """HTCondor job flavours and their respective runtime in seconds."""
+
+KUBERNETES_MEMORY_UNITS = ["E", "P", "T", "G", "M", "K"]
+"""Kubernetes valid memory units"""
+
+KUBERNETES_MEMORY_FORMAT = r"\d+[{}]i?\b".format("".join(KUBERNETES_MEMORY_UNITS))
+"""Kubernetes valid memory format regular expression e.g. Ki, M , Gi, G, etc."""
 
 REANA_RUNTIME_KUBERNETES_KEEP_ALIVE_JOBS_WITH_STATUSES = os.getenv(
     "REANA_RUNTIME_KUBERNETES_KEEP_ALIVE_JOBS_WITH_STATUSES", ""