+++ reana_commons/job_utils.py
@@ -8,6 +8,9 @@
 """REANA-Commons job utils."""
 
 import base64
+import re
+
+from reana_commons.config import KUBERNETES_MEMORY_FORMAT
 
 
 def serialise_job_command(command):
@@ -18,3 +21,8 @@ def serialise_job_command(command):
 def deserialise_job_command(command):
     """Deserialise job commands received through REST API."""
     return base64.b64decode(command).decode("utf-8")
+
+
+def validate_kubernetes_memory(memory):
+    """Verify that provided value matches the Kubernetes memory format."""
+    return re.match(KUBERNETES_MEMORY_FORMAT, memory) is not None