+++ reana_commons/serial.py
@@ -13,6 +13,7 @@ from copy import deepcopy
 from string import Template
 
 from jsonschema import ValidationError, validate
+from reana_commons.config import KUBERNETES_MEMORY_FORMAT
 from reana_commons.utils import check_htcondor_max_runtime
 
 serial_workflow_schema = {
@@ -62,6 +63,7 @@ serial_workflow_schema = {
                         "$id": "#/properties/steps/properties/kubernetes_memory_limit",
                         "type": "string",
                         "default": "",
+                        "pattern": KUBERNETES_MEMORY_FORMAT,
                     },
                     "voms_proxy": {
                         "$id": "#/properties/steps/properties/voms_proxy",